requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
simple-salesforce==1.12.9
google-generativeai==0.8.6
Faker==40.8.0
chromadb==1.5.2
